Heroku link does not work
I built a compatibility-based "FriendFinder" application -- basically a dating app. This full-stack site will take in results from a user's survey, then compare their answers with those from other users. The app will then display the name and picture of the user with the best overall match.

The app uses express to handle routing.

This app is deployed in Heroku, so follow this link to use it: https://serene-earth-21209.herokuapp.com/